naukari_automation/requirements.txt
selenium>=4.18.0
webdriver-manager>=4.0.0
google-generativeai>=0.8.0
PyPDF2>=3.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
